Fix ruff rule F401: remove unused imports (#31)
* Remove unused imports

* Move fixture to conftest.py

* Enable rule F401

* Remove one more import
