+++ docs/conf.py
@@ -22,7 +22,6 @@
 import os
 import sys
 import shutil
-import sphinx_rtd_theme
 print(os.path.abspath('./shap'))
 sys.path.insert(0, os.path.abspath('..'))
