+++ shap/_explanation.py
@@ -2,8 +2,6 @@
 import pandas as pd
 import numpy as np
 import scipy as sp
-import sys
-import warnings
 import copy
 import operator
 import sklearn
