+++ shap/benchmark/framework.py
@@ -1,11 +1,7 @@
 import matplotlib.pyplot as plt 
-import pandas as pd 
 import numpy as np 
 import itertools as it 
-import sklearn 
-import shap 
-from sklearn.model_selection import train_test_split
-from shap.utils import safe_isinstance, MaskedModel
+from shap.utils import safe_isinstance
 from . import perturbation
 
 def update(model, attributions, X, y, masker, sort_order, perturbation_method, scores):
