+++ shap/benchmark/measures.py
@@ -1,6 +1,5 @@
 import numpy as np
 from tqdm.autonotebook import tqdm
-import gc
 import warnings
 import sklearn.utils
