+++ shap/plots/_embedding.py
@@ -1,9 +1,7 @@
-import numpy as np
 import sklearn
 import warnings
 try:
     import matplotlib.pyplot as pl
-    import matplotlib
 except ImportError:
     warnings.warn("matplotlib could not be loaded!")
     pass
