+++ shap/plots/_group_difference.py
@@ -2,7 +2,6 @@ import numpy as np
 import warnings
 try:
     import matplotlib.pyplot as pl
-    import matplotlib
 except ImportError:
     warnings.warn("matplotlib could not be loaded!")
     pass
