+++ shap/plots/_monitoring.py
@@ -3,7 +3,6 @@ import scipy
 import warnings
 try:
     import matplotlib.pyplot as pl
-    import matplotlib
 except ImportError:
     warnings.warn("matplotlib could not be loaded!")
     pass
