+++ shap/utils/_general.py
@@ -1,12 +1,9 @@
 import re
-import pandas as pd
 import numpy as np
 import scipy as sp
-from scipy.spatial.distance import pdist
 import sys
 import warnings
 import sklearn
-import importlib
 import copy
 from contextlib import contextmanager
 import sys, os
