+++ tests/plots/test_summary.py
@@ -1,4 +1,3 @@
-import warnings
 import matplotlib.pyplot as plt
 import numpy as np
 import pytest
