+++ tests/plots/test_waterfall.py
@@ -1,7 +1,6 @@
 import matplotlib.pyplot as plt
 import pytest
 import shap
-from .utils import explainer # (pytest fixture do not remove) pylint: disable=unused-import
 
 @pytest.mark.mpl_image_compare(tolerance=3)
 def test_waterfall(explainer): # pylint: disable=redefined-outer-name
